handle json questions
