+++ test.cjs
@@ -1,4 +1,4 @@
-import interrogation from './index.cjs';
+const interrogation = require('./index.cjs');
 
 // eslint-disable-next-line max-len
 /* interrogation.getAnswersFor(['question1? ', 'question2? ', ['question3', ['choice1', 'choice2', 'choice3']], 'question4? ']).then((value) => {
@@ -9,6 +9,8 @@ import interrogation from './index.cjs';
   console.log(`value : ${value}`);
 }); */
 
-// console.log(interrogation.encodeQuestions('json-question.json'));
+console.log(interrogation.encodeQuestions('json-question.json'));
 
-console.log(interrogation.decodeQuestions('question.int'));
+/* interrogation.getAnswersForJson('question.int').then((value) => {
+  console.log(value);
+}); */
